+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 -   ITop configuration variables
 
+### Changed
+
+-   Removed modules: bicycle-parkings, parking-zones, public-toilets
+
 ## [2.5.10] - 2023-04-03
 
 -   No changelog